+++ app/components/common/Button/BaseButton.tsx
@@ -1,0 +1,33 @@
+
+import React, { MouseEvent } from 'react';
+import { ButtonType } from './type';
+
+const BaseButton = ({
+  children,
+  onClick,
+  className,
+  theme = '',
+  type = 'button',
+  disabled = false,
+}: ButtonType) => {
+  const clickFn = (e: MouseEvent<HTMLButtonElement>) => {
+    e.stopPropagation();
+    e.preventDefault();
+    if (typeof onClick === 'function') {
+      onClick(e);
+    }
+  };
+
+  return (
+    <button
+      className={`component ${theme} ${disabled ? 'disabled' : ''} ${className}`}
+      disabled={disabled}
+      type={type}
+      onClick={clickFn}
+    >
+      { children }
+    </button>
+  );
+};
+
+export default BaseButton;
