+++ app/components/common/Button/PrimaryButton.tsx
@@ -1,0 +1,17 @@
+
+import React from 'react';
+
+import { FlexibleButtonProps } from './type';
+import BaseButton from './BaseButton';
+
+const PrimaryButton = ({
+  className,
+  ...restProps
+}: FlexibleButtonProps) => (
+  <BaseButton
+    {...restProps}
+    className={`primaryButton ${className}`}
+  />
+);
+
+export default PrimaryButton;
