+++ app/components/common/Button/button.css
@@ -5,6 +5,7 @@
   &.iconButton {
     color: var(--white);
     cursor: pointer;
+    -webkit-tap-highlight-color: transparent;
   }
 
   &.primaryButton,
